change all customer request in entity/dto relation that has double data type to Big Decimal and add condition to temiRate by age and car series

+++ src/main/java/com/app/smartdrive/api/services/customer/CustomerInscExtendService.java
@@ -3,6 +3,7 @@ package com.app.smartdrive.api.services.customer;
 import com.app.smartdrive.api.entities.customer.CustomerInscAssets;
 import com.app.smartdrive.api.entities.customer.CustomerInscExtend;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CustomerInscExtendService {
@@ -10,7 +11,7 @@ public interface CustomerInscExtendService {
             Long[] cuexIds,
             CustomerInscAssets cias,
             Long entityId,
-            Double currentPrice
+            BigDecimal currentPrice
     );
 
     public void deleteAllCustomerInscExtendInCustomerRequest(Long creqEntityId);
